fix: return type

+++ internal/domain/values/reports.go
@@ -4,6 +4,6 @@ import "time"
 
 type RecurringRevenue struct {
 	Period time.Time `json:"period"`
-	Total  int64     `json:"total"`
+	Total  float64   `json:"total"`
 	Type   string    `json:"type"`
 }
